[fix] 마이페이지 관련 api User 도메인 -> Onboarding 도메인으로 이동
